+++ torch/csrc/utils/cpp_stacktraces.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <torch/csrc/Export.h>
+
+namespace torch {
+TORCH_API bool get_cpp_stacktraces_enabled();
+}
